+++ tests/cortex/test_cortex_commands.py
@@ -227,3 +227,21 @@ def test_cortex_translate_file(_mock_cortex_result, runner):
         assert_successful_result_message(
             result, expected_msg="John ma samochód. Samochód Johna jest niebieski."
         )
+
+
+@mock.patch("snowflake.cli.plugins.cortex.commands.SEARCH_COMMAND_ENABLED", new=False)
+def test_if_search_raises_exception_for_312(runner, snapshot):
+
+    result = runner.invoke(
+        [
+            "cortex",
+            "search",
+            "parrot",
+            "--service",
+            "test_service",
+            "--columns",
+            "region",
+        ]
+    )
+    assert result.exit_code == 1
+    assert result.output == snapshot