+++ tests_integration/test_cortex.py
@@ -1,3 +1,5 @@
+import sys
+
 import pytest
 
 from tests_integration.conftest import TEST_DIR
@@ -5,6 +7,35 @@ from tests_integration.testing_utils.assertions.test_result_assertions import (
     assert_successful_result_message,
     assert_that_result_is_successful,
 )
+
+
+@pytest.mark.integration
+@pytest.mark.skip  # TODO: when we have Cortex activated on test account, unskip this and setup part in sql script
+@pytest.mark.skipif(
+    sys.version_info >= (3, 12),
+    reason="Snowflake Python API currently does not support Python 3.12 and greater",
+)
+def test_cortex_search(runner):
+    result = runner.invoke_with_connection_json(
+        [
+            "cortex",
+            "search",
+            "parrot",
+            "--service",
+            "test_service",
+            "--columns",
+            "region",
+        ]
+    )
+
+    expected_result = [
+        {
+            "": "This parrot is no more! It has ceased to be! It`s expired and gone to meet its maker!",
+            "region": "Flying Circus",
+        }
+    ]
+    assert result.exit_code == 0
+    assert result.json == expected_result
 
 
 @pytest.mark.integration